Overhauled and simplified functionality. Mobile first with Bootstrap 3

+++ application/controllers/book.php
@@ -5,96 +5,108 @@ class Book extends CI_Controller {
 	public function __construct()
 	{
 		parent::__construct();
-
+		lockdown();
 		$this->load->library('form_validation');
-		$this->load->model('book_model');
 		$this->load->model('chapter_model');
+		$this->load->model('article_model');
+		$this->load->model('user_model');
+		$this->load->model('book_model');
+		
 
 		$this->form_validation->set_error_delimiters('<div class="form-error text-danger">', '</div>');
 	}
 
 
 	public function index(){
-		$this->all();
+		$data['title'] = 'My Books';
+		$data['books'] = $this->book_model->getAllBooks();
+		$this->template->inject('book/list', $data);
 	}
 
 	
 	public function add(){
 
-		// Breadcrumbs
-		$this->breadcrumb->clear();
-		$this->breadcrumb->add_crumb('Books', base_url('book/index')); // this will be a link
-		$this->breadcrumb->add_crumb('Add New', base_url('book/add')); // this will be a link
-		$data['bread'] = $this->breadcrumb->output();
-
 		$data['title'] = 'Add new book';
 
 		// Form rules
 		$this->form_validation->set_rules('name', 'book name', 'trim|required|xss_clean');
-		$this->form_validation->set_rules('description', 'book description', 'trim|required|xss_clean');
+		$this->form_validation->set_rules('description', 'book description', 'trim|xss_clean');
 
 		if(!$this->form_validation->run()){
 			$this->template->inject('book/add', $data);
 		} else {
 			
 			if($this->book_model->addNewBook($this->input->post())){
-				redirect(base_url('book/all?status=success'));
+				redirect(base_url('book?status=book_added'));
 			} else {
+				$data['msgBox'] = $this->arena->renderMsgBox('An error occured while creating this book. Please try again later', 'Ooops...');
 				$this->template->inject('book/add', $data);
 			}
 		}
 	}
 
-	/*Shows all books in db*/
-	public function all(){
-
-		// Breadcrumbs
-		$this->breadcrumb->clear();
-		$this->breadcrumb->add_crumb('All Books', base_url('book/index')); // this will be a link
-		$data['bread'] = $this->breadcrumb->output();
-
-
-		$data['title'] = 'My Books';
-		$data['books'] = $this->book_model->getAllBooks();
-
-		$this->template->inject('book/list', $data);
-	}
 
 	public function edit($bookID){
 
 		$data['title'] = 'Edit Book';
-		$data['book'] = $this->book_model->findBook($bookID);
+		$data['book'] = $this->book_model->find($bookID);
 
-
-		// Breadcrumbs
-		$this->breadcrumb->clear();
-		$this->breadcrumb->add_crumb('Books', base_url('book/index'));
-		$this->breadcrumb->add_crumb($data['book']['name'], base_url('book/index'));
-		$this->breadcrumb->add_crumb('Edit', '#');
-		$data['bread'] = $this->breadcrumb->output();
-		
 
 		$this->form_validation->set_rules('name', 'book name', 'trim|required|xss_clean');
-		$this->form_validation->set_rules('description', 'book description', 'trim|required|xss_clean');
+		$this->form_validation->set_rules('description', 'book description', 'trim|xss_clean');
 
 		if(!$this->form_validation->run()){
 			$this->template->inject('book/edit', $data);
 		} else {
 			// Save book
 			if($this->book_model->editBook($this->input->post())){
-				redirect(base_url('book/all?status=success'));
+				redirect(base_url('books?status=edited'));
 			} else {
-				redirect(base_url('book/all?status=failed'));
+				redirect(base_url('books?status=edit_failed'));
 			}
 		}
 	}
 
 	public function delete($bookID){
-		if($this->book_model->deleteBook($bookID)){
-			redirect(base_url('book/all?status=success'));
+
+		if(!$this->input->post('posted')){
+			$data['title'] = 'Confirm Action';
+			$data['book'] = $this->book_model->find($bookID);
+			$this->template->inject('book/confirm-delete', $data);
 		} else {
-			redirect(base_url('book/all?status=failed'));
+
+			// var_dump($_POST);
+
+			// exit();
+
+			if($this->input->post('confirm_action')==true){
+				if($this->book_model->deleteBook($bookID)){
+					redirect(base_url('books?status=deleted'));
+				} else {
+					redirect(base_url('books?status=delete_failed'));
+				}
+			} else {
+				redirect(base_url('books?status=delete_cancelled'));
+			}
+ 
 		}
+
+	}
+
+	public function download($bookID){
+
+		$data['book'] = $this->book_model->findBook($bookID);
+
+		$chapters = $this->chapter_model->getAllChapters($bookID);
+		foreach ($chapters as $key => $chapter) {
+			$chapters[$key]['articles'] = $this->article_model->getChapterArticles($chapter['id']);
+		}
+
+
+		$data['book']['chapters'] = $chapters;
+
+		$this->load->view('book/export', $data);
+
 	}
 
 }
